requirements.txt
Flask
mysql-connector-python
boto3
